Updated MosaicAlias
Fixed layout issue, moved unit test into namespace file.
